Second page progress!
Finally happy with a studio for the unboxing. It's a little
claustrophobic but that discomfort is good for a sense of place on the
web.

Moving forward: OrbitControls need to be fixed, code needs to be cleaned
(second.js is currently a graveyard of failed attempts and versions of
today), and CSS Animation needs to be implemented for the dummy 'buddha'
cube.